feat(user-review): Edit data type of field Rating

+++ internal/model/review.go
@@ -6,7 +6,7 @@ type Review struct {
 	ID          string    `json:"id" gorm:"primaryKey;size:255"`
 	BikeId      string    `json:"bike_id" gorm:"size:255"`
 	UserId      string    `json:"user_id" gorm:"size:255"`
-	Rating      float32   `json:"rating"`
+	Rating      int       `json:"rating"`
 	Description string    `json:"description"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
